move onPanZoom to React component
